Funcionalidad guardar compra separada del componente
